requirements.txt
# Core application dependencies
# Install with: pip install -r requirements.txt

anthropic>=0.40.0
kafka-python>=2.0.2
praw>=7.7.0
tweepy>=4.14.0
playwright>=1.44.0
SQLAlchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
redis>=5.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
tenacity>=8.3.0
